# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation Moon Group",
        "description": "API Documentation for Moon Group",
        "version": "1.0.0"
    },
    "paths": {
...
        "/moon-group-backend/public/api/news": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get all news",
                "description": "Get all news",
                "operationId": "71a7fb4708318b291d9c91001159c1ab",
                "parameters": [
                    {
                        "parameter": "all",
                        "name": "all",
                        "in": "query",
                        "description": "Get all news",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "parameter": "page",
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "parameter": "per_page",
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "parameter": "sort",
                        "name": "sort",
                        "in": "query",
                        "description": "Sort by column",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "parameter": "direction",
                        "name": "direction",
                        "in": "query",
                        "description": "Sort direction",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsCollection"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Create news",
                "description": "Create news",
                "operationId": "bc3ef079968f000d9f9644a7dcaefe14",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreNewsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/moon-group-backend/public/api/news/{id}": {
            "get": {
                "tags": [
                    "News"
                ],
                "summary": "Get news by id",
                "description": "Get news by id",
                "operationId": "c5292e5597035fb65856c83ab5e3c832",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "News id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Noticia no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "News"
                ],
                "summary": "Delete news",
                "description": "Delete news",
                "operationId": "b80cbb87dc1e7caa1c844e4a064c5d0d",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "News id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Noticia no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/moon-group-backend/public/api/news/update/{id}": {
            "post": {
                "tags": [
                    "News"
                ],
                "summary": "Update news",
                "description": "Update news",
                "operationId": "70cdf4f5e3339c15e642cf06557c43e7",
                "parameters": [
                    {
                        "parameter": "id",
                        "name": "id",
                        "in": "path",
                        "description": "News id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateNewsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NewsResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Noticia no encontrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "example": "https://develop.garzasoft.com/moon-group-backend/public/api/path?page=1"
                    },
                    "last": {
                        "type": "string",
                        "example": "https://develop.garzasoft.com/moon-group-backend/public/api/path?page=4"
                    },
                    "prev": {
                        "type": "string",
                        "example": "null"
                    },
                    "next": {
                        "type": "string",
                        "example": "https://develop.garzasoft.com/moon-group-backend/public/api/path?page=2"
                    }
                },
                "type": "object"
            },
            "PaginationMetaLinks": {
                "properties": {
                    "url": {
                        "type": "string",
                        "example": "https://develop.garzasoft.com/moon-group-backend/public/api/path?page=1"
                    },
                    "label": {
                        "type": "string",
                        "example": "1"
                    },
                    "active": {
                        "type": "boolean",
                        "example": "true"
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": "1"
                    },
                    "from": {
                        "type": "integer",
                        "example": "1"
                    },
                    "last_page": {
                        "type": "integer",
                        "example": "4"
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationMetaLinks"
                    },
                    "path": {
                        "type": "string",
                        "example": "https://develop.garzasoft.com/moon-group-backend/public/api/path"
                    },
                    "per_page": {
                        "type": "integer",
                        "example": "15"
                    },
                    "to": {
                        "type": "integer",
                        "example": "15"
                    },
                    "total": {
                        "type": "integer",
                        "example": "60"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "The pagination must be an integer."
                    }
                },
                "type": "object"
            },
            "Unauthenticated": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
...
            "StoreNewsRequest": {
                "title": "StoreNewsRequest",
                "required": [
                    "title",
                    "date",
                    "introduction",
                    "description",
                    "image"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Title"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2021-09-01"
                    },
                    "introduction": {
                        "type": "string",
                        "example": "Introduction"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description"
                    },
                    "image": {
                        "type": "file",
                        "format": "binary"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
            "UpdateNewsRequest": {
                "title": "Update News Request",
                "required": [
                    "title",
                    "date",
                    "introduction",
                    "description",
                    "image"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Title"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2021-09-01"
                    },
                    "introduction": {
                        "type": "string",
                        "example": "Introduction"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description"
                    },
                    "image": {
                        "type": "file",
                        "format": "binary"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
            "NewsResource": {
                "title": "NewsResource",
                "description": "News resource",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "type": "string",
                        "example": "Title"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-08-19"
                    },
                    "introduction": {
                        "type": "string",
                        "example": "Introduction"
                    },
                    "description": {
                        "type": "string",
                        "example": "Description"
                    },
                    "image": {
                        "type": "string",
                        "example": "Image"
                    },
                    "active": {
                        "type": "boolean",
                        "example": "1"
                    },
                    "category_id": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "NewsCollection": {
                "title": "NewsCollection",
                "description": "News resource collection",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NewsResource"
                        }
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationLinks"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}